Factor out Track setting fields from variant track settings, reuse fo… (#24)
* Factor out Track setting fields from variant track settings, reuse for enhancer track settings

* clean up
